% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbbplotR-package.R, R/geom_cbb_conferences.R,
%   R/geom_cbb_headshots.R, R/geom_cbb_teams.R
\docType{package}
\name{cbbplotR-package}
\alias{cbbplotR}
\alias{cbbplotR-package}
\alias{GeomCBBconference}
\alias{GeomCBBheads}
\alias{GeomCBBteam}
\title{cbbplotR: CBB Logo Plots in 'ggplot2'}
\format{
An object of class \code{GeomCBBconference} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{GeomCBBheads} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{GeomCBBteam} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomCBBconference

GeomCBBheads

GeomCBBteam
}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of functions to visualize college basketball team and conference logos + player headshots in 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andreweatherman/cbbplotR}
  \item \url{https://cbbplotr.aweatherman.com}
  \item \url{https://andreweatherman.github.io/cbbplotR/}
  \item Report bugs at \url{https://github.com/andreweatherman/cbbplotR/issues}
}

}
\author{
\strong{Maintainer}: Andrew Weatherman \email{andrew@aweatherman.com}

}
\keyword{datasets}
\keyword{internal}
