% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cbbplotR_clear_cache}
\alias{cbbplotR_clear_cache}
\title{Clear cbbplotR Cache}
\usage{
cbbplotR_clear_cache()
}
\value{
Invisibly \code{NULL}
}
\description{
This function clears the memoised cache of all functions memoised by \code{cbbplotR}.
}
\examples{
cbbplotR_clear_cache()
}
