% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.R
\name{element_cbb_teams}
\alias{element_cbb_teams}
\alias{element_cbb_headshots}
\alias{element_cbb_conferences}
\alias{element_path}
\title{Theme elements for ggplot2}
\usage{
element_cbb_teams(
  alpha = NULL,
  colour = NA,
  hjust = NULL,
  vjust = NULL,
  color = NULL,
  size = 0.5
)

element_cbb_headshots(
  player_id_column,
  alpha = NULL,
  colour = NA,
  hjust = NULL,
  vjust = NULL,
  color = NULL,
  size = 1.5
)

element_cbb_conferences(
  alpha = NULL,
  colour = NA,
  hjust = NULL,
  vjust = NULL,
  color = NULL,
  size = 0.5
)

element_path(...)
}
\arguments{
\item{alpha}{The transparency level \link{0-1}.}

\item{hjust, vjust}{The horizontal and vertical adjustment respectively \link{0-1}.}

\item{color}{The image will be colorized with this color. Use the special character \code{"b/w"} to set it to black and white.}

\item{size}{The output grob size in \code{cm}.}

\item{...}{
  Arguments passed on to \code{\link[ggpath:element_path]{ggpath::element_path}}
  \describe{
    \item{\code{colour,color}}{The image will be colorized with this color. Use the
special character \code{"b/w"} to set it to black and white. For more information
on valid color names in ggplot2 see
\url{https://ggplot2.tidyverse.org/articles/ggplot2-specs.html?q=colour#colour-and-fill}.}
    \item{\code{hjust,vjust}}{The horizontal and vertical adjustment respectively.
Must be a numerical value between 0 and 1.}
    \item{\code{angle}}{The angle of the element as a numerical value between 0° and 360°.}
  }}
}
\value{
An S3 object of class \code{element}.
}
\description{
These functions enable images in non-data parts of ggplot2 layers.
}
\details{
\itemize{
\item \code{element_cbb_logo()}: draws CBB team logos instead of their names.
\item \code{element_cbb_conferences()}: draws CBB conference logos instead of their names.
\item \code{element_cbb_headshot()}: draws CBB player headshots instead of their ESPN player IDs.
\item \code{element_path()}: draws images from valid image URLs instead of the URL.
}
}
