% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{element_cbb}
\alias{element_cbb_teams}
\alias{element_cbb_headshots}
\alias{element_cbb_conferences}
\alias{element_grob.element_cbb_teams}
\alias{element_grob.element_cbb_headshots}
\alias{element_grob.element_cbb_conferences}
\title{Theme elements for ggplot2 to Visualize College Basketball Elements}
\usage{
element_cbb_teams(
  alpha = NULL,
  colour = NA,
  hjust = NULL,
  vjust = NULL,
  color = NULL,
  size = 0.5
)

element_cbb_headshots(
  player_id_column,
  alpha = NULL,
  colour = NA,
  hjust = NULL,
  vjust = NULL,
  color = NULL,
  size = 1.5
)

element_cbb_conferences(
  alpha = NULL,
  colour = NA,
  hjust = NULL,
  vjust = NULL,
  color = NULL,
  size = 0.5,
  logo_type = "normal"
)

\method{element_grob}{element_cbb_teams}(
  element,
  label = "",
  x = NULL,
  y = NULL,
  alpha = NULL,
  colour = NULL,
  hjust = 0.5,
  vjust = 0.5,
  size = NULL,
  ...
)

\method{element_grob}{element_cbb_headshots}(
  element,
  label = "",
  x = NULL,
  y = NULL,
  alpha = NULL,
  colour = NULL,
  hjust = 0.5,
  vjust = 0.5,
  size = NULL,
  ...
)

\method{element_grob}{element_cbb_conferences}(
  element,
  label = "",
  x = NULL,
  y = NULL,
  alpha = NULL,
  colour = NULL,
  hjust = 0.5,
  vjust = 0.5,
  size = NULL,
  ...
)
}
\arguments{
\item{alpha}{The transparency level \link{0-1}.}

\item{hjust, vjust}{Horizontal and vertical adjustment \link{0-1}.}

\item{color}{(Optional) Colorization for the images; use "b/w" for black and white.}

\item{size}{Size of the output grob in cm.}

\item{player_id_column}{Column name for player IDs, applicable for \code{element_cbb_headshots}.}

\item{logo_type}{Type of logo to use ('normal' or 'wordmark'), applicable for \code{element_cbb_conferences}.}

\item{...}{
  Arguments passed on to \code{\link[ggpath:element_path]{ggpath::element_path}}
  \describe{
    \item{\code{colour,color}}{The image will be colorized with this color. Use the
special character \code{"b/w"} to set it to black and white. For more information
on valid color names in ggplot2 see
\url{https://ggplot2.tidyverse.org/articles/ggplot2-specs.html?q=colour#colour-and-fill}.}
    \item{\code{hjust,vjust}}{The horizontal and vertical adjustment respectively.
Must be a numerical value between 0 and 1.}
    \item{\code{angle}}{The angle of the element as a numerical value between 0° and 360°.}
  }}
}
\value{
An S3 object of class \code{element} appropriate for use in ggplot2 theme settings.
}
\description{
These functions enable the use of images in non-data parts of ggplot2 layers, specifically for
visualizing elements related to college basketball.
}
\section{Functions}{

\itemize{
\item \code{element_cbb_teams()}: Replaces text with team logos in ggplot2 themes.
\item \code{element_cbb_conferences()}: Replaces text with conference logos in ggplot2 themes.
\item \code{element_cbb_headshots()}: Replaces ESPN player IDs with player headshots in ggplot2 themes.
}
}

\examples{
# Usage within a ggplot theme setting:
ggplot(data, aes(x, y)) +
  theme(axis.text.x = element_cbb_teams(size = 1.0))
}
