% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cbb_conferences.R
\name{geom_cbb_conferences}
\alias{geom_cbb_conferences}
\title{Visualize College Basketball Conference Logos in ggplot2}
\usage{
geom_cbb_conferences(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = TRUE,
  logo_type = "normal",
  highlight_conferences = NULL,
  highlight_method = "alpha",
  highlight_alpha = 0.5
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{ggplot2::aes()} or \code{ggplot2::aes_()}.
If specified and \code{inherit.aes = TRUE} (the default), it is combined with the default
mapping at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. If \code{NULL}, the default, the data is
inherited from the plot data as specified in the call to \code{ggplot()}.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a
position adjustment function.}

\item{...}{Other arguments passed on to \code{layer()}.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a warning.
If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{Logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE} never includes,
and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining
with them. This is most useful for helper functions that define both data and
aesthetics and shouldn't inherit behaviour from the default plot specification,
e.g. \code{borders()}.}

\item{logo_type}{Character specifying the type of logos to plot (\code{"normal"} or \code{"wordmark"}).}

\item{highlight_conferences}{Vector of conference names to highlight.}

\item{highlight_method}{Method of highlighting (\code{"alpha"}, \code{"color"}, or \code{"both"}).}

\item{highlight_alpha}{Alpha value for highlighted conferences.}
}
\value{
A ggplot2 layer that can be added to a plot created with \code{ggplot()}.
}
\description{
\code{geom_cbb_conferences} creates a ggplot2 layer that plots college basketball conference logos
in place of points. The layer requires \code{x} and \code{y} aesthetics and a valid conference name.
The \code{width} parameter can be used to adjust the size of the logos. Logos can be depicted as
either \code{normal} or \code{wordmark} by setting the \code{logo_type} parameter.
}
\section{Aesthetics}{

This geom requires the following aesthetics:
\itemize{
\item{\strong{x}}{ - The x-coordinate for the logo's position.}
\item{\strong{y}}{ - The y-coordinate for the logo's position.}
\item{\strong{conference}}{ - The name of the conference, which should be a valid name from \code{cbbdata}.}
}
}

