% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_espn_players.R
\name{get_espn_players}
\alias{get_espn_players}
\title{Get ESPN Player IDs for a team}
\usage{
get_espn_players(team)
}
\arguments{
\item{team}{Team name}
}
\description{
This grab player IDs from ESPN to be used as headshot column
}
\examples{
\donttest {
cbbplotR::get_espn_players('Duke')
}
}
