% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\name{ggplot_cbb_logo_title}
\alias{ggplot_cbb_logo_title}
\title{Add a team logo in-line with plot title with \code{ggplot2}}
\usage{
ggplot_cbb_logo_title(
  value,
  type = c("team", "conference", "headshot"),
  size = 5,
  hjust = 1,
  vjust = 1,
  position = c(1, 1)
)
}
\arguments{
\item{size}{The size of the plotted logo}

\item{hjust}{The horizontal position of the logo}

\item{vjust}{The vertical position of the logo}

\item{position}{Something...}

\item{team}{The team whose logo will be plotted. Must match team names from
\code{cbbdata}.}
}
\description{
Add a team logo in-line with plot title with \code{ggplot2}
}
