% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-utils.R
\name{gt_cbb_conferences}
\alias{gt_cbb_conferences}
\title{Render in Conference Wordmarks in \code{gt} Tables}
\usage{
gt_cbb_conferences(
  data,
  conf_column,
  logo_column = "conf_logo",
  logo_height = 20,
  include_name = FALSE
)
}
\arguments{
\item{data}{Pass through your plotting data}

\item{conf_column}{Indicate which column contains your \code{team} variable}

\item{logo_column}{Indicates the name of the HTML column}

\item{logo_height}{The height of the logo or wordmark in the HTML string}

\item{include_name}{Should the conference name be included in the column?
Defaults to FALSE as wordmarks are distinctive.}
}
\value{
Returns data with an appended HTML column.
}
\description{
Appends a column with HTML that will render in conference wordmarks and
(optionally) names for table plotting.
}
\details{
I really hate putting conference logos and names in separate columns,
and this helper function will add conference wordmarks and create
the HTML code to render in wordmarks and columns in the same column for tables.
If you are using \code{gt} to plot, you must use \code{fmt_markdown({logo_column})} to
render in the HTML. Set \code{logo_column} to your \code{conf} variable if you wish
to overwrite your \code{conf} column with the logo HTML (recommended but not default).
}
\examples{
\dontrun{try(cbd_torvik_ratings(year=2023) \%>\% head() \%>\% select(team, conf) \%>\%
cbbdata::cbd_gt_logos(team, team) \%>\% gt::gt() \%>\% gt::fmt_markdown(team))}

}
