% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-utils.R
\name{gt_cbb_logo_title}
\alias{gt_cbb_logo_title}
\title{Add a Team Logo and Title/Subtitle to a \code{gt} Table.}
\usage{
gt_cbb_logo_title(
  title,
  subtitle,
  value = NULL,
  logo_link = NULL,
  type = c("team", "conference", "headshot"),
  title_font_size = 24,
  title_font_weight = "bold",
  title_lineheight = 0.5,
  subtitle_font_size = 16,
  subtitle_font_weight = "normal",
  subtitle_lineheight = 0.5,
  logo_height = 65
)
}
\arguments{
\item{title}{The \code{gt} table title}

\item{subtitle}{The \code{gt} table subtitle}

\item{logo_link}{If you want to plot a different image, pass a valid URL link to it.}

\item{title_font_size}{The size of the title font (in px).}

\item{title_font_weight}{The weight of the title (normal or bold).}

\item{title_lineheight}{The line-height of the title.}

\item{subtitle_font_size}{The size of the subtitle font (in px).}

\item{subtitle_font_weight}{The weight of the subtitle (normal or bold).}

\item{subtitle_lineheight}{The line-height of the subtitle.}

\item{logo_height}{The height of the logo (in px).}

\item{team}{A valid team name from \code{cbbdata}.}
}
\description{
This requires wrapping the function around \code{html} and passing
through to \code{tab_header} in \code{gt}. See function example.
}
\examples{
\donttest {
library(gt)
library(dplyr)
library(cbbplotR)

gt(head(mtcars)) \%>\%
tab_header(title = html(gt_cbb_logo_title('Test title', 'Test subtitle', 'Duke')))

}
}
