% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-utils.R
\name{gt_theme_pl}
\alias{gt_theme_pl}
\title{Premier League \code{gt} Table Theme}
\usage{
gt_theme_pl(gt_object, ...)
}
\arguments{
\item{gt_object}{An existing gt table object of class \code{gt_tbl}}

\item{...}{Optional additional arguments to \code{gt::table_options()}}
}
\value{
Returns data with an appended HTML column.
}
\description{
A theme for styling \code{gt} tables similar to the Premier League
official table.
}
