% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_cbb}
\alias{scale_color_cbb_teams}
\alias{scale_colour_cbb_teams}
\alias{scale_fill_cbb_teams}
\alias{scale_color_cbb_conferences}
\alias{scale_colour_cbb_conferences}
\alias{scale_fill_cbb_conferences}
\title{Color and Fill Scales for College Basketball Teams and Conferences}
\usage{
scale_color_cbb_teams(
  type = c("primary", "secondary"),
  values = NULL,
  ...,
  aesthetics = "colour",
  breaks = ggplot2::waiver(),
  na.value = "grey50",
  guide = NULL,
  alpha = NA
)

scale_colour_cbb_teams(
  type = c("primary", "secondary"),
  values = NULL,
  ...,
  aesthetics = "colour",
  breaks = ggplot2::waiver(),
  na.value = "grey50",
  guide = NULL,
  alpha = NA
)

scale_fill_cbb_teams(
  type = c("primary", "secondary"),
  values = NULL,
  ...,
  aesthetics = "fill",
  breaks = ggplot2::waiver(),
  na.value = "grey50",
  guide = NULL,
  alpha = NA
)

scale_color_cbb_conferences(
  type = c("primary", "secondary"),
  values = NULL,
  ...,
  aesthetics = "colour",
  breaks = ggplot2::waiver(),
  na.value = "grey50",
  guide = NULL,
  alpha = NA
)

scale_colour_cbb_conferences(
  type = c("primary", "secondary"),
  values = NULL,
  ...,
  aesthetics = "colour",
  breaks = ggplot2::waiver(),
  na.value = "grey50",
  guide = NULL,
  alpha = NA
)

scale_fill_cbb_conferences(
  type = c("primary", "secondary"),
  values = NULL,
  ...,
  aesthetics = "fill",
  breaks = ggplot2::waiver(),
  na.value = "grey50",
  guide = NULL,
  alpha = NA
)
}
\arguments{
\item{type}{The type of color scale, either \code{"primary"} or \code{"secondary"}.}

\item{values}{A named list of colors to be used if overriding the default team or conference colors.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:discrete_scale]{discrete_scale}}
  \describe{
    \item{\code{palette}}{A palette function that when called with a single integer
argument (the number of levels in the scale) returns the values that
they should take (e.g., \code{\link[scales:pal_hue]{scales::pal_hue()}}).}
    \item{\code{breaks}}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}
    \item{\code{limits}}{One of:
\itemize{
\item \code{NULL} to use the default scale values
\item A character vector that defines possible values of the scale and their
order
\item A function that accepts the existing (automatic) values and returns
new ones. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}
    \item{\code{drop}}{Should unused factor levels be omitted from the scale?
The default, \code{TRUE}, uses the levels that appear in the data;
\code{FALSE} uses all the levels in the factor.}
    \item{\code{na.translate}}{Unlike continuous scales, discrete scales can easily show
missing values, and do so by default. If you want to remove missing values
from a discrete scale, specify \code{na.translate = FALSE}.}
    \item{\code{na.value}}{If \code{na.translate = TRUE}, what aesthetic value should the
missing values be displayed as? Does not apply to position scales
where \code{NA} is always placed at the far right.}
    \item{\code{aesthetics}}{The names of the aesthetics that this scale works with.}
    \item{\code{scale_name}}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} The name of the scale
that should be used for error messages associated with this scale.}
    \item{\code{labels}}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item An expression vector (must be the same length as breaks). See ?plotmath for details.
\item A function that takes the breaks as input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}
    \item{\code{call}}{The \code{call} used to construct the scale for reporting messages.}
    \item{\code{super}}{The super class to use for the constructed scale}
  }}

\item{aesthetics}{The ggplot2 aesthetic to apply the scale to, defaults to \code{"colour"} or \code{"fill"}.}

\item{breaks}{A set of breaks for the scale, defaults to \code{ggplot2::waiver()}.}

\item{na.value}{The color to use for missing values.}

\item{guide}{A guide for the scale, such as a legend.}

\item{alpha}{Transparency level \link{0-1}, applied to all colors if not \code{NA}.}
}
\value{
A ggplot2 color or fill scale.
}
\description{
These functions provide color and fill scales for ggplot2 based on college basketball team and conference colors.
}
\section{Functions}{

\itemize{
\item \code{scale_color_cbb_teams()}: Provides color scales for college basketball teams.
\item \code{scale_fill_cbb_teams()}: Provides fill scales for college basketball teams.
\item \code{scale_color_cbb_conferences()}: Provides color scales for college basketball conferences.
\item \code{scale_fill_cbb_conferences()}: Provides fill scales for college basketball conferences.
}
}

